services/document-analysis-service/requirements.txt
flask
python-dotenv
gunicorn
python-dateutil # For date calculations (e.g., expiry checks)
